backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
drf-spectacular>=0.27,<1
celery>=5.3,<6
psycopg[binary]>=3.1,<4
gunicorn>=22,<23
pytest>=8,<9
pytest-django>=4.9,<5
